requirements.txt
django==5.2.9
djangorestframework==3.16.1
dj-database-url==2.3.0
psycopg2-binary==2.9.9

